Alter Marks constructor behavior so that properties may be passed in
As we move towards having marks be something that can be created by
dispatching a store action, or loaded from a flat JSON export, we need
a way to reconstitute that externally-specified mark as the relevant
sort of primitive.

This PR accomplishes the first step of that: it alters marks so that
they accept a configuration object specifying, among other things,
the properties for the mark that is being constructed. By passing these
properties in we allow all marks to be set at one location.

The next step will be to permit marks like Rect to accept their own
configuration object, which could be merged with these defaults and
then sent on to Marks; that configuration object could (will) originate
from the redux store, via a dispatch action.