Add marks to primitive store via action and parse values to signals
One action is dispatched for each mark addition: That action will be used by
the primitives reducer to create the mark object within the primitives store.
Mark properties are mapped to their signal equivalents on insertion, rather
than ex post facto on init; this is so that the signals reducer can handle the
same mark add action to register the appropriate signals at the same time the
mark is registered with the store.